rename react query queries folder
