+++ packages/manager/src/features/LoadBalancers/LoadBalancerDetail/LoadBalancerSummary.tsx
@@ -6,7 +6,7 @@ import { Paper } from 'src/components/Paper';
 import { Stack } from 'src/components/Stack';
 import { Typography } from 'src/components/Typography';
 import { IPAddress } from 'src/features/Linodes/LinodesLanding/IPAddress';
-import { useLoadBalancerQuery } from 'src/queries/aglb/loadbalancers';
+import { useLoadBalancerQuery } from 'src/queries/aclb/loadbalancers';
 // import { useRegionsQuery } from 'src/queries/regions';
 
 import { Ports } from '../LoadBalancerLanding/Ports';
